remove useAuth.js and move the useAuth() hook into AuthProvider
