+++ src/components/MemoDetail.jsx
@@ -1,4 +1,4 @@
-import { useAuth } from "../hooks/useAuth";
+import { useAuth } from "./AuthProvider";
 
 export default function MemoDetail(props) {
   const { isLoggedIn } = useAuth();
